This fix allows loading external libs with buster.js tests
When loading external libraries with async-loader, there seems a error when the loader tries to inject the script code into the header. It tries to inject before the "first_script", which is ok in cases where async-loader is in fact the first script. However, in case of buster.js tests it seems the case that there is nothing in the header (first_script.parentNode). I added a little fix that checks if that node is avail, and if not simply appends the new script to the head by el.appendChild.